✅[chore]: 가격 관련 상수 Double 타입으로 변경

+++ OpenMarket/OpenMarket/Model/ProductInformation.swift
@@ -13,9 +13,9 @@ struct ProductInformation {
     let name: String
     let thumbnail: String
     let currency: Currency
-    let price: Int
-    let barginPrice: Int
-    let discountedPrice: Int
+    let price: Double
+    let barginPrice: Double
+    let discountedPrice: Double
     let stock: Int
     let createdAt: Date
     let issuedAt: Date
